% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Agrivoltaics Statistics Shiny App}
\usage{
launch_app(port = NULL, browser = TRUE)
}
\arguments{
\item{port}{Port number (default: random available port)}

\item{browser}{Open in browser automatically (default: TRUE)}
}
\description{
Opens the agrivoltaics statistical analysis interface in your browser.
}
\examples{
\dontrun{
launch_app()
}

}
